refactor: Move friend and group in model directory
