+++ src/core/coreav.cpp
@@ -21,8 +21,8 @@
 #include "core.h"
 #include "coreav.h"
 #include "src/audio/audio.h"
-#include "src/friend.h"
-#include "src/group.h"
+#include "src/model/friend.h"
+#include "src/model/group.h"
 #include "src/persistence/settings.h"
 #include "src/video/corevideosource.h"
 #include "src/video/videoframe.h"
