+++ src/video/groupnetcamview.cpp
@@ -20,7 +20,7 @@
 #include "groupnetcamview.h"
 #include "src/audio/audio.h"
 #include "src/core/core.h"
-#include "src/friend.h"
+#include "src/model/friend.h"
 #include "src/friendlist.h"
 #include "src/nexus.h"
 #include "src/persistence/profile.h"
