+++ src/video/videosurface.cpp
@@ -19,7 +19,7 @@
 
 #include "videosurface.h"
 #include "src/core/core.h"
-#include "src/friend.h"
+#include "src/model/friend.h"
 #include "src/friendlist.h"
 #include "src/persistence/settings.h"
 #include "src/video/videoframe.h"
