+++ src/widget/contentdialog.cpp
@@ -33,9 +33,9 @@
 #include "style.h"
 #include "widget.h"
 #include "src/core/core.h"
-#include "src/friend.h"
+#include "src/model/friend.h"
 #include "src/friendlist.h"
-#include "src/group.h"
+#include "src/model/group.h"
 #include "src/grouplist.h"
 #include "src/persistence/settings.h"
 #include "src/widget/form/chatform.h"
