+++ src/widget/form/genericchatform.cpp
@@ -22,9 +22,9 @@
 #include "src/chatlog/chatlog.h"
 #include "src/chatlog/content/timestamp.h"
 #include "src/core/core.h"
-#include "src/friend.h"
+#include "src/model/friend.h"
 #include "src/friendlist.h"
-#include "src/group.h"
+#include "src/model/group.h"
 #include "src/grouplist.h"
 #include "src/persistence/settings.h"
 #include "src/persistence/smileypack.h"
