+++ src/widget/form/groupchatform.cpp
@@ -27,9 +27,9 @@
 #include "tabcompleter.h"
 #include "src/core/core.h"
 #include "src/core/coreav.h"
-#include "src/friend.h"
+#include "src/model/friend.h"
 #include "src/friendlist.h"
-#include "src/group.h"
+#include "src/model/group.h"
 #include "src/video/groupnetcamview.h"
 #include "src/widget/flowlayout.h"
 #include "src/widget/form/chatform.h"
