+++ src/widget/groupwidget.cpp
@@ -31,9 +31,9 @@
 #include "maskablepixmapwidget.h"
 #include "form/groupchatform.h"
 #include "src/core/core.h"
-#include "src/friend.h"
+#include "src/model/friend.h"
 #include "src/friendlist.h"
-#include "src/group.h"
+#include "src/model/group.h"
 #include "src/grouplist.h"
 #include "src/widget/friendwidget.h"
 #include "src/widget/style.h"
